+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 2.0.1
+- Find port in wwise global settings
+
 ## 2.0.0
-- migrate to dart 3.0
+- Migrate to dart 3.0
 
 ## 1.0.1
-- add basic WObject types
+- Add basic WObject types
 
 ## 1.0.0